% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_quantity.R
\name{set_quantity}
\alias{set_quantity}
\title{Sets quantity of token}
\usage{
set_quantity(token, amount, side)
}
\arguments{
\item{amount}{enum in BUSD}

\item{side}{string BUY or SELL}

\item{symbol}{string}
}
\value{
R object
}
\description{
Sets quantity of token
}
\examples{
\dontrun{
set_filters('BTC', amount = 20, side = 'BUY')
}
}
